Write a function that prints the binary representation of a number

+++ bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
 
 void print_binary(unsigned long int n)
 {
-	int i = 0, y, k, size;
+	unsigned long int i = 0, y, k, size;
 
 	if (n == 0)
 		printf("0");
@@ -18,7 +18,7 @@ void print_binary(unsigned long int n)
 		i++;
 		y = y >> 1;
 	}
-	
+
 	for (size = i - 1; size >= 0; size--)
 	{
 		k = n >> size;
